+++ CHANGELOG.md
@@ -18,23 +18,30 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 * Add geo shape query filter ([#319](https://github.com/opensearch-project/dashboards-maps/pull/319))
 * Add shape filter UI button ([#329](https://github.com/opensearch-project/dashboards-maps/pull/329))
 * Add tooltip to draw shape filter ([#330](https://github.com/opensearch-project/dashboards-maps/pull/330))
+
 ### Enhancements
 * Enhance layer visibility status display ([#299](https://github.com/opensearch-project/dashboards-maps/pull/299))
 * Introduce disable tooltip on hover property ([#313](https://github.com/opensearch-project/dashboards-maps/pull/313))
 * Update tooltip behavior change ([#317](https://github.com/opensearch-project/dashboards-maps/pull/317))
 * Update max supported layer count ([#332](https://github.com/opensearch-project/dashboards-maps/pull/332))
 * BWC for document layer label textType ([#340](https://github.com/opensearch-project/dashboards-maps/pull/340))
+
 ### Bug Fixes
 * Fix property value undefined check ([#276](https://github.com/opensearch-project/dashboards-maps/pull/276))
 * Show scroll bar when panel height reaches container bottom ([#295](https://github.com/opensearch-project/dashboards-maps/pull/295))
 * fix: fixed filters not reset when index pattern changed ([#234](https://github.com/opensearch-project/dashboards-maps/pull/234))
 * Add custom layer visibility config to render ([#297](https://github.com/opensearch-project/dashboards-maps/pull/297))
 * Fix color picker component issue ([#305](https://github.com/opensearch-project/dashboards-maps/pull/305))
-* fix: layer filter setting been reset unexpectedly ([#327](https://github.com/opensearch-project/dashboards-maps/commits/2.x))
+* fix: layer filter setting been reset unexpectedly ([#327](https://github.com/opensearch-project/dashboards-maps/pull/327))
+* Fix data query in dashboard mode when enable around map filter ([#339](https://github.com/opensearch-project/dashboards-maps/pull/339))
+
 ### Infrastructure
 * Add CHANGELOG ([#342](https://github.com/opensearch-project/dashboards-maps/pull/342))
+
 ### Documentation
+
 ### Maintenance
+
 ### Refactoring
 * Move zoom and coordinates as separate component ([#309](https://github.com/opensearch-project/dashboards-maps/pull/309))
 * Move coordinates to footer ([#315](https://github.com/opensearch-project/dashboards-maps/pull/315))